chg: usr: adjust command output so it continues onto the next line
files affected:
- linux/lib/html-functions.sh
- linux/etc/default.conf

Some unix commands generate output with very long lines of output.
The program was previously cutting off all text beyond the 250'th
character.  The full output width of commands are now shown in the
reports.

This change introduces a new variable in the config file, namely
CFG_TEXTWIDTH, which is used to specify the width of the window
that the program should use before continuing the output onto the
next line.  This variable will have a new option (-w) defined in an
upcoming change that will allow this to be set from the command line.
The default setting for this variable is set to what was the previously
hard-coded value of 74 characters wide.  Note that this value never had
any effect on the executed internal linux command output, as many of
them generated output longer than 74 characters.  Where this value had
the most effect was in the section headers in the generated ASCII file.