chg: usr: Record the paragraph title and executed command in the ASCII file
files affected:
- linux/lib/html-functions.sh

Until now, the paragraph title and the executed command are shown under
each heading in the generated HTML file, but not in the ASCII file.
This change puts the paragraph title and the command or function
called, in the .txt file.  So that they can each be easily found in
the report, the commands begin with the prefix "###".

This extends the meaning of the CFG_STINLINE variable (-L option), to
now also include recording the command in the ASCII file by default. If
the variable is set to "no" or the -L option is specified, recording
of the command will be turned off in both generated reports.