chg: usr: Command option additions (-z, -v, -w)
files affected:
- linux/etc/default.conf
- linux/cfg2html-linux.sh

Code was recently added to the program to gather information regarding
ZFS filesystems and VMware.  However, there was no means to turn off
incorporating these sections into the report.  New variables have been
added to support their inclusion (by default), or exclusion.  By using
the -z and -v options, we can now selectively turn off collecting
information for the ZFS filesystem and VMware sections, respectively.

The new command-line option, -w allows a report width value to be
provided to the program at runtime.  This overrides the default and
previously hardcoded value of 74.  Depending on your X-window screen
size in which you view the report, a value of 300 to 350 was found
to be quite useful.