chg: show all ext2-4 partitions in display_ext_fs_param function
files affected:
- linux/lib/linux-functions.sh

If the lsblk command is in the PATH, then execute it to find
all partitions that are ext2, 3, or 4, be them mounted or not,
and display the filesystem information in turn for each of them.
If lsblk is not available, then as before this change, only details
about the mounted filesystems is provided in the report.