chg: usr: added column option for redhat rpm commands
files affected:
- linux/cfg2html-linux-sh

CentOS generated a very long list of RPM package information, so the
column command was added in case the -w option was used.

column cmd  was added to the selinux command as well.

The column command was removed from ldconfig, as it made the output
very messy looking.