chg: fix: adjustment for null character in /proc/<pid>/cmdline
files affected:
- linux/lib/linux-functions.sh

function affected:
- topFDhandles

It was found that the null character \000, or (^@) was appended to
the text in /proc/<pid>/cmdline in some cases. For each affected
process id, this caused an error message in the journal of the form:

BASH_ERR: internal_warning [ (11161)"bash" -> ...  CMD=$(cat /proc/9673/cmdline)" ] sh: command substitution: ignored null byte in input

To resolve the issue, the "cat" command was substituted with "strings".